requirements.txt
flask 
gunicorn
-e .
pandas
scikit-learn
pandas 
numpy
dill
evidently

pyYAML

